Create hurdleBoot_code.r
This R script fits via ordinary bootstrap a Hurdle model to a count outcome on a set of pre-specified covariates and an intervention assignment binary indicator in both its zero and positive part, and returns a vector of inferences (i.e. point estimate, 95% confidence interval and p-value) for the intervention effect parameter that are averaged across the covariate distribution.

The code requires as user input specification of: 

--  an individual-patient level rectangular data-frame, whereby the count outcome is labelled "Outcome" and the intervention assignment indicator is labelled "Intervention";
--  the GLM family used to represent the outcome-covariates relation in the positive part of the Hurdle model (defaulting in the code to Negative Binomial);
--  the method of calculation of the bootstrap confidence interval for the intervention effect (defaulting in the code to biased-corrected accelerated);
--  names of predictors (and interactions thereof) to populate the positive and zero parts of the Hurdle regression equation;
--  the size of each bootstrap sample (defaulting in the code to 5000).

Lastly an assumption is made in tehg code that there is no interaction between the intervention assignment indicator and any of the covariates in the Hurdle model.